Deduplicate agents fake process tests
