+++ packages/agents/src/prd-generator.test.ts
@@ -1,4 +1,3 @@
-import { EventEmitter } from 'node:events';
 import { afterEach, describe, expect, it, vi } from 'vitest';
 
 const mockSpawn = vi.hoisted(() => vi.fn());
@@ -13,37 +12,7 @@ vi.mock('./health-check', () => ({
 }));
 
 import { buildPrdPrompt, enhancePrdDraft, extractPrd } from './prd-generator';
-
-function createFakeProc() {
-  const proc = new EventEmitter() as EventEmitter & {
-    stdout: EventEmitter;
-    stderr: EventEmitter;
-    stdin: { write: (chunk: string) => boolean; end: () => void };
-  };
-  proc.stdout = new EventEmitter();
-  proc.stderr = new EventEmitter();
-  const stdinWrites: string[] = [];
-  let stdinEnded = false;
-  proc.stdin = {
-    write: (chunk: string) => {
-      stdinWrites.push(chunk);
-      return true;
-    },
-    end: () => {
-      stdinEnded = true;
-    },
-  };
-  return {
-    proc,
-    stdinWrites,
-    isStdinEnded: () => stdinEnded,
-    close: (code: number, options?: { stdout?: string; stderr?: string }) => {
-      if (options?.stdout) proc.stdout.emit('data', options.stdout);
-      if (options?.stderr) proc.stderr.emit('data', options.stderr);
-      proc.emit('close', code);
-    },
-  };
-}
+import { createFakeProc } from './test-fake-proc';
 
 describe('enhancePrdDraft', () => {
   afterEach(() => {
@@ -51,7 +20,7 @@ describe('enhancePrdDraft', () => {
   });
 
   it('surfaces Claude stdout result when the CLI exits non-zero with empty stderr', async () => {
-    const fake = createFakeProc();
+    const fake = createFakeProc({ captureStdin: true });
     mockSpawn.mockReturnValueOnce(fake.proc);
 
     const promise = enhancePrdDraft({
@@ -81,7 +50,7 @@ describe('enhancePrdDraft', () => {
   });
 
   it('passes explicit Claude model without adding thinking tokens for low effort', async () => {
-    const fake = createFakeProc();
+    const fake = createFakeProc({ captureStdin: true });
     mockSpawn.mockReturnValueOnce(fake.proc);
 
     const promise = enhancePrdDraft({
@@ -122,7 +91,7 @@ describe('enhancePrdDraft', () => {
   });
 
   it('uses Codex CLI when selected', async () => {
-    const fake = createFakeProc();
+    const fake = createFakeProc({ captureStdin: true });
     mockSpawn.mockReturnValueOnce(fake.proc);
 
     const promise = enhancePrdDraft({
