+++ packages/agents/src/skill-rewriter.test.ts
@@ -1,4 +1,3 @@
-import { EventEmitter } from 'node:events';
 import { afterEach, describe, expect, it, vi } from 'vitest';
 
 const mockSpawn = vi.hoisted(() => vi.fn());
@@ -17,40 +16,10 @@ import {
   extractRewrittenSkill,
   rewriteSkillDraft,
 } from './skill-rewriter';
+import { createFakeProc } from './test-fake-proc';
 
 const VALID_PLAN_SKILL =
   '---\nname: plan-generation\n---\nUse {{USER_PROMPT}} for {{THREAD_ID}} and emit {{OUTPUT_SCHEMA}}.';
-
-function createFakeProc() {
-  const proc = new EventEmitter() as EventEmitter & {
-    stdout: EventEmitter;
-    stderr: EventEmitter;
-    stdin: { write: (chunk: string) => boolean; end: () => void };
-  };
-  proc.stdout = new EventEmitter();
-  proc.stderr = new EventEmitter();
-  const stdinWrites: string[] = [];
-  let stdinEnded = false;
-  proc.stdin = {
-    write: (chunk: string) => {
-      stdinWrites.push(chunk);
-      return true;
-    },
-    end: () => {
-      stdinEnded = true;
-    },
-  };
-  return {
-    proc,
-    stdinWrites,
-    isStdinEnded: () => stdinEnded,
-    close: (code: number, options?: { stdout?: string; stderr?: string }) => {
-      if (options?.stdout) proc.stdout.emit('data', options.stdout);
-      if (options?.stderr) proc.stderr.emit('data', options.stderr);
-      proc.emit('close', code);
-    },
-  };
-}
 
 describe('rewriteSkillDraft', () => {
   afterEach(() => {
@@ -132,7 +101,7 @@ describe('rewriteSkillDraft', () => {
   });
 
   it('pipes the rewrite prompt through Claude stdin and returns valid skill content', async () => {
-    const fake = createFakeProc();
+    const fake = createFakeProc({ captureStdin: true });
     mockSpawn.mockReturnValueOnce(fake.proc);
 
     const promise = rewriteSkillDraft({
@@ -167,7 +136,7 @@ describe('rewriteSkillDraft', () => {
   });
 
   it('rejects rewritten content that loses required slots', async () => {
-    const fake = createFakeProc();
+    const fake = createFakeProc({ captureStdin: true });
     mockSpawn.mockReturnValueOnce(fake.proc);
 
     const promise = rewriteSkillDraft({
@@ -194,7 +163,7 @@ describe('rewriteSkillDraft', () => {
   });
 
   it('rejects rewritten content that invents unsupported runtime slots', async () => {
-    const fake = createFakeProc();
+    const fake = createFakeProc({ captureStdin: true });
     mockSpawn.mockReturnValueOnce(fake.proc);
 
     const promise = rewriteSkillDraft({
